requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
scikit-learn
mlflow
pyaml
fastapi
uvicorn

#-e .